% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_geometry_comunas}
\alias{census_geometry_comunas}
\title{Chilean municipalities polygons as simple features}
\format{
A data frame with 345 rows and 2 variables:
\describe{
  \item{code}{municipality code}
  \item{geometry}{municipality geometry}
}
}
\source{
\url{http://www.ine.cl/}
}
\usage{
census_geometry_comunas
}
\description{
An sf object containing the chilean municipalities.
}
\keyword{datasets}
