% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinim.R
\name{get_sinim}
\alias{get_sinim}
\title{Get a SINIM variable data in a specific year as a data frame}
\usage{
get_sinim(var, year, moncorr = T, idgeo = F, geometry = F, region,
  province, comuna)
}
\arguments{
\item{var}{Variable code}

\item{year}{Year}

\item{moncorr}{A logical value indicating the use of monetary correction}

\item{idgeo}{A logical value to add province and region columns}

\item{geometry}{A logical value to add geographical features}

\item{region}{Regional subsetting variable}

\item{province}{Province subsetting variable}

\item{comuna}{Comuna subsetting variable}
}
\value{
data frame with data for the requested variable over time with optional geometry
}
\description{
Get a SINIM variable data in a specific year as a data frame
}
\examples{
get_sinim(880, 2015)
get_sinim(880, 2015:2017)
get_sinim(c(880, 882), 2015)
get_sinim(c(880, 882), 2015:2017)
get_sinim(c(880, 882), 2015, idgeo=TRUE)
get_sinim(882, 2015, geometry=TRUE)
get_sinim(882, 2015, region="02")
}
