% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinim_cats.R
\name{get_sinim_cats}
\alias{get_sinim_cats}
\title{Get SINIM data categories}
\usage{
get_sinim_cats()
}
\value{
Returns a data frame with available data in SINIM's webpage
}
\description{
Get SINIM data categories
}
\examples{
get_sinim_cats()
}
