% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinim_var_name.R
\name{get_sinim_var_name}
\alias{get_sinim_var_name}
\title{Return SINIM variable name using their internal codes}
\usage{
get_sinim_var_name(var)
}
\arguments{
\item{var}{Variable code}
}
\value{
A vector of requested variable names
}
\description{
Return SINIM variable name using their internal codes
}
\examples{
get_sinim_var_name(880)
get_sinim_var_name(c(880, 882))
}
