% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinim_vars.R
\name{get_sinim_vars}
\alias{get_sinim_vars}
\title{List available variables for a category}
\usage{
get_sinim_vars(catn)
}
\arguments{
\item{catn}{category number}
}
\value{
data frame with variables and values
}
\description{
List available variables for a category
}
\examples{
get_sinim_vars(47)
}
