% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsinimcategories.R
\name{getsinimcategories}
\alias{getsinimcategories}
\title{Get categories of data in SINIM}
\usage{
getsinimcategories()
}
\value{
data frame with available data
}
\description{
Get categories of data in SINIM
}
\examples{
getsinimcategories()
}
