% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsinimr.R
\name{getsinimr}
\alias{getsinimr}
\title{Get a SINIM variable data in a specific year as a data frame}
\usage{
getsinimr(var, year, moncorr = T)
}
\arguments{
\item{var}{variable code}

\item{year}{year}

\item{moncorr}{A logical value indicating the use of monetary correction}
}
\value{
data frame with data for the requested variable over time
}
\description{
Get a SINIM variable data in a specific year as a data frame
}
\examples{
getsinimr(880, 2015)
getsinimr(c(880,882), 2015)
}
