% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsinimryears.R
\name{getsinimryears}
\alias{getsinimryears}
\title{Get data for a variable by multiple years as a data frame}
\usage{
getsinimryears(var, years)
}
\arguments{
\item{var}{variable code}

\item{years}{years}
}
\value{
data frame with data for the requested variable over time
}
\description{
Get data for a variable by multiple years as a data frame
}
\examples{
getsinimr(880, c(2015,2014,2013))
}
