% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsinimvariables.R
\name{getsinimvariables}
\alias{getsinimvariables}
\title{List available variables for a category}
\usage{
getsinimvariables(catn)
}
\arguments{
\item{catn}{category number}
}
\value{
data frame with variables and values
}
\description{
List available variables for a category
}
\examples{
getsinimvariables(47)
}
