% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{id_geo_census}
\alias{id_geo_census}
\title{Chilean administrative divisions}
\format{
A data frame with 345 rows and 7 variables:
\describe{
  \item{code}{municipality code}
  \item{municipality}{municipality name}
  \item{code.reg}{region code}
  \item{nom.reg}{region name}
  \item{code.prov}{province code}
  \item{nom.prov}{province name}
  \item{auc}{yes/no consolidated urban area}
}
}
\source{
\url{http://www.ine.cl/}
}
\usage{
id_geo_census
}
\description{
A dataset containing the chilean administrative units.
}
\keyword{datasets}
