% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idgeocensus}
\alias{idgeocensus}
\title{Geographical identifier}
\format{A data frame with 246 rows and 6 variables:
\describe{
  \item{CODE.REG}{Region code}
  \item{NOM.REG}{Region name}
  \item{CODE.PROV}{Provincia code}
  \item{NOM.PROV}{Provincia name}
  \item{CODE}{Municipality code}
  \item{MUNICIPALITY}{Municipality name}
}}
\source{
\url{http://www.censo2017.cl}
}
\usage{
idgeocensus
}
\description{
A dataset containing the geographical units of the
chilean 2017 census.
}
\keyword{datasets}
