% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm.R
\name{mm}
\alias{mm}
\title{Return a abbreviated value of the "miles de millones" format}
\usage{
mm(x)
}
\arguments{
\item{x}{Numeric value to convert}
}
\value{
A character value in short format
}
\description{
Return a abbreviated value of the "miles de millones" format
}
\examples{
mm(5807587000)
}
