% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_sinim_vars.R
\name{search_sinim_vars}
\alias{search_sinim_vars}
\title{Search for a SINIM variable and if no keyword is provided returns all variables}
\usage{
search_sinim_vars(keyword)
}
\arguments{
\item{keyword}{keyword}
}
\value{
data frame with results
}
\description{
Search for a SINIM variable and if no keyword is provided returns all variables
}
\examples{
search_sinim_vars("ingresos propios")
}
