% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchsinimvar.R
\name{searchsinimvar}
\alias{searchsinimvar}
\title{Search for a SINIM variable and if no keyword is provided returns all variables}
\usage{
searchsinimvar(keyword)
}
\arguments{
\item{keyword}{keyword}
}
\value{
data frame with results
}
\description{
Search for a SINIM variable and if no keyword is provided returns all variables
}
\examples{
searchsinimvar()
searchsinimvar("ingresos propios")
}
