% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinimr.R
\docType{package}
\name{sinimr}
\alias{sinimr}
\alias{sinimr-package}
\title{R library for SINIM}
\description{
Access SINIM data and view it as data frames.
}
\author{
Roberto Salas \email{rosalas@uc.cl}
}
\keyword{package}
