% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinim.R
\name{get_sinim}
\alias{get_sinim}
\title{Get a SINIM variable data in a specific year as a data frame}
\usage{
get_sinim(
  var,
  year,
  moncorr = T,
  truevalue = F,
  idgeo = F,
  geometry = F,
  region,
  provincia,
  comuna,
  auc = F,
  unit = "comunas"
)
}
\arguments{
\item{var}{Variable code}

\item{year}{Year}

\item{moncorr}{A logical value indicating the use of monetary correction}

\item{truevalue}{A logical value indicating the use of converted values}

\item{idgeo}{A logical value to add provincia and region columns}

\item{geometry}{A logical value to add geographical features}

\item{region}{Region subsetting variable}

\item{provincia}{Provincia subsetting variable}

\item{comuna}{Comuna subsetting variable}

\item{auc}{A logical value to retrieve only AUC features}

\item{unit}{Use "comunas" or "limites"}
}
\value{
data frame for the requested variable over time with optional geometry
}
\description{
Get a SINIM variable data in a specific year as a data frame
}
\examples{
get_sinim(880, 2015)
get_sinim(880, 2015:2017)
get_sinim(c(880, 882), 2015)
get_sinim(c(880, 882), 2015:2017)
get_sinim(c(880, 882), 2015, idgeo=TRUE)
get_sinim(882, 2015, geometry=TRUE)
get_sinim(882, 2015, region="02")
}
