% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsinimcategories.R
\name{getsinimcategories}
\alias{getsinimcategories}
\title{Get SINIM data categories}
\usage{
getsinimcategories()
}
\value{
Returns a data frame with available data in SINIM's webpage
}
\description{
Get SINIM data categories
}
\examples{
getsinimcategories()
}
