% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsinimr.R
\name{getsinimr}
\alias{getsinimr}
\title{Get catch data for a region as a dataframe}
\usage{
getsinimr(var, year)
}
\arguments{
\item{year}{year id}

\item{variable}{variable type}
}
\value{
data frame with data for the requested variable over time
}
\description{
Get catch data for a region as a dataframe
}
\examples{
getsinimr(880, 2015)
getsinimr(c(880,1023), 2015)
}
