% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchsinimvar.R
\name{searchsinimvar}
\alias{searchsinimvar}
\title{Search a SINIM variable}
\usage{
searchsinimvar(keyword)
}
\arguments{
\item{keyword}{keyword}
}
\value{
data frame with results
}
\description{
Search a SINIM variable
}
\examples{
searchsinimvar("ingresos propios")
}
